+++ main.h
@@ -22,6 +22,6 @@ int character(va_list arg);
 int string(va_list arg);
 int percent(va_list arg);
 int decimal(va_list arg);
-void printer(int x);
+void printer(long int x);
 int pch(char c);
 #endif
